requirements.txt
flask==2.3.3
twilio==8.10.0
streamlit==1.28.1
reportlab==4.0.4
apscheduler==3.10.4
sqlite3
pandas==2.1.1
plotly==5.17.0
python-dotenv==1.0.0